fixed????? item collection
one of the pages that was fully broken now works if you just. set it as "bugged" which does NOTHING on its own. and then have an offset of 0 (which at values other than 0 actually does something in conjunction to bugged) i cannot explain why this is and it does not make sense. works though.